requirements.txt
click==8.1.3
dash==2.9.3
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.3
imageio==2.28.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-loader==0.2
MarkupSafe==2.1.2
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
scikit-image==0.20.0
scipy==1.9.1
six==1.16.0
tenacity==8.2.2
tifffile==2023.4.12
tzdata==2023.3
Werkzeug==2.2.3
zipp==3.15.0
